handlecount: rename handles package
The `handles` package name actually clashed with a ton of other places
that had variables called `handles`. This commit renames the package to
`handlecount`.
