+++ receiver/hostmetricsreceiver/internal/scraper/processscraper/internal/handlecount/handles_windows.go
@@ -4,7 +4,7 @@
 //go:build windows
 // +build windows
 
-package handles // import "github.com/open-telemetry/opentelemetry-collector-contrib/receiver/hostmetricsreceiver/internal/scraper/processscraper/internal/handles"
+package handlecount // import "github.com/open-telemetry/opentelemetry-collector-contrib/receiver/hostmetricsreceiver/internal/scraper/processscraper/internal/handlecount"
 
 import (
 	"errors"
