+++ receiver/hostmetricsreceiver/internal/scraper/processscraper/internal/handlecount/handles_windows_test.go
@@ -4,7 +4,7 @@
 //go:build windows
 // +build windows
 
-package handles
+package handlecount
 
 import (
 	"errors"
